[SAM_E51_CNANO] Regenerated demo for same51n_low_power_with_oled_c_click
[DOCS] Tested same51n_getting_started demo and updated readme, json files

M32MM-1192

+++ apps/sam_e51_cnano/same51n_low_power_with_oled_c_click/firmware/src/config/sam_e51_cnano/system/time/src/sys_time.c
@@ -166,18 +166,18 @@ static void SYS_TIME_HwTimerCompareUpdate(void)
     {
         if (tmrActive->relativeTimePending > SYS_TIME_HW_COUNTER_HALF_PERIOD)
         {
-            nextHwCounterValue = counterObj->hwTimerCurrentValue + SYS_TIME_HW_COUNTER_HALF_PERIOD;
+            nextHwCounterValue = (uint64_t)counterObj->hwTimerCurrentValue + SYS_TIME_HW_COUNTER_HALF_PERIOD;
         }
         else
         {
             /* Use a non-volatile intermediate to prevent dual volatile access in single statement */
             uint32_t relativeTimePending = tmrActive->relativeTimePending;
-            nextHwCounterValue = counterObj->hwTimerCurrentValue + relativeTimePending;
+            nextHwCounterValue = (uint64_t)counterObj->hwTimerCurrentValue + relativeTimePending;
         }
     }
     else
     {
-        nextHwCounterValue = counterObj->hwTimerCurrentValue + SYS_TIME_HW_COUNTER_HALF_PERIOD;
+        nextHwCounterValue = (uint64_t)counterObj->hwTimerCurrentValue + SYS_TIME_HW_COUNTER_HALF_PERIOD;
     }
 
     currHwCounterValue = counterObj->timePlib->timerCounterGet();
